started vgg16 compression of dense layers
